- GeoIP code had a hardcoded path eventhough there was a config field for this - wrapped all the loading and querying around if blocks and exception handlers to make it all better
Update issue 20
The code in geoip.rb has been made more robust, it seems the error from geoip that u found is 
untrappable since its a seg fault, so I wrapped it in an if to make sure the file exists first
before attempting to load it.

I've also made the actual country() call return nil on failure such as this so if you have else 
blocks in your case statements they'll cleanly handle the failure of missing/corrupt data.

Thanks for your help in finding this one.